cleaning and adding readme
