+++ samples/java-protobuf-tracing/README.md
@@ -1,29 +1,5 @@
-# tracing
+## This example show how to create Spans by the users
  
-
- Make a sync call to generate a factorial 
- make an async call calling to the component from an Action that makes factorial
-
-## Designing
-
-While designing your service it is useful to read [designing services](https://docs.kalix.io/developing/development-process-proto.html)
-
-
-## Developing
-
-This project has a bare-bones skeleton service ready to go, but in order to adapt and
-extend it, it may be useful to read up on [developing services](https://docs.kalix.io/services/)
-and in particular the [Java section](https://docs.kalix.io/java-protobuf/index.html)
-
-
-## Building
-
-You can use Maven to build your project, which will also take care of
-generating code based on the `.proto` definitions:
-
-```shell
-mvn compile
-```
 
 
 ## Running Locally
@@ -38,29 +14,33 @@ mvn kalix:runAll
 ```
 
 This command will start your Kalix service and a companion Kalix Runtime as configured in [docker-compose.yml](./docker-compose.yml) file.
+This will also start a Jaeger service to which the services above will push the traces. 
+You can find Jaeger at `localhost:16686`
 
-> Note: if you're looking to use Google Pub/Sub, see comments inside [docker-compose.yml](./docker-compose.yml)
-> on how to enable a Google Pub/Sub emulator that Kalix Runtime will connect to.
 
-With both the Kalix Runtime and your service running, any defined endpoints should be available at `http://localhost:9000`. In addition to the defined gRPC interface, each method has a corresponding HTTP endpoint. Unless configured otherwise (see [Transcoding HTTP](https://docs.kalix.io/java-protobuf/writing-grpc-descriptors-protobuf.html#_transcoding_http)), this endpoint accepts POST requests at the path `/[package].[entity name]/[method]`. For example, using `curl`:
-
-```shell
-> curl -XPOST -H "Content-Type: application/json" localhost:9000/com.example.CounterService/GetCurrentCounter -d '{"counterId": "foo"}'
-The command handler for `GetCurrentCounter` is not implemented, yet
-```
-
+With both the Kalix Runtime and your service running, any defined endpoints should be available at `http://localhost:9000`. In addition to the defined gRPC interface, each method has a corresponding HTTP endpoint. Unless configured otherwise (see [Transcoding HTTP](https://docs.kalix.io/java-protobuf/writing-grpc-descriptors-protobuf.html#_transcoding_http)), this endpoint accepts POST requests at the path `/[package].[entity name]/[method]`.
 For example, using [`grpcurl`](https://github.com/fullstorydev/grpcurl):
 
 ```shell
-> grpcurl -plaintext -d '{"counterId": "foo"}' localhost:9000 com.example.CounterService/GetCurrentCounter 
-ERROR:
-  Code: Unknown
-  Message: The command handler for `GetCurrentCounter` is not implemented, yet
+grpcurl -plaintext localhost:9000  com.example.Controller/CallAsyncEndpoint                                 
 ```
+produces
+```
+{
+  "message": "{\n  \"userId\": 1,\n  \"id\": 1,\n  \"title\": \"sunt aut facere repellat provident occaecati excepturi optio reprehenderit\",\n  \"body\": \"quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto\"\n}"
+}
+```
+or 
 
-> Note: The failure is to be expected if you have not yet provided an implementation of `GetCurrentCounter` in
-> your entity.
-
+```shell
+grpcurl -plaintext localhost:9000  com.example.Controller/CallAsyncEndpoint  
+```
+produces
+```
+{
+  "message": "{\n  \"userId\": 1,\n  \"id\": 1,\n  \"title\": \"sunt aut facere repellat provident occaecati excepturi optio reprehenderit\",\n  \"body\": \"quia et suscipit\\nsuscipit recusandae consequuntur expedita et cum\\nreprehenderit molestiae ut ut quas totam\\nnostrum rerum est autem sunt rem eveniet architecto\"\n}"
+}
+```
 
 ## Deploying
