--- samples/java-protobuf-tracing/src/main/proto/com/example/domain/factorial_domain.proto
@@ -1,9 +1,0 @@
-syntax = "proto3";
-
-package com.example.domain;
-
-option java_outer_classname = "FactorialDomain";
-
-message FactorialState {
-  string value = 1;
-}
