+++ samples/java-protobuf-tracing/src/main/resources/application.conf
@@ -1,0 +1,1 @@
+kalix.telemetry.tracing.collector-endpoint="http://localhost:4317"
